moved PageContext out of backgrounds
