+++ src/components/LinksBar.tsx
@@ -1,4 +1,4 @@
-import { usePageContext } from '../backgrounds/PageContext'
+import { usePageContext } from '../PageContext'
 import { defaultTransition } from '../App'
 import { Screen } from '../utils/screen'
 import { pinks } from './Button'
