+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { PageProvider } from './backgrounds/PageContext';
+import { PageProvider } from './PageContext';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
